refactor: convert content-generator homepage to server component with auth-based routing
- Convert page.tsx from client component to async server component
- Split UI into PublicHomepage and Workspace client components
- Add lucide-react dependency for icon support
- Update middleware to allow unauthenticated access to homepage (/)
- Show Workspace for authenticated users, PublicHomepage for guests
- Update pnpm-lock.yaml with lucide-react and peer dependency resolution
