+++ apps/content-generator/middleware.ts
@@ -3,8 +3,8 @@ import { createAuthMiddleware } from "@asafarim/auth/middleware";
 const portalUrl = process.env.PORTAL_URL || "https://portal-qa.asafarim.com";
 
 export const middleware = createAuthMiddleware({
-  // Registered users only — only the health check is public
-  publicRoutes: ["/api/health"],
+  // Public routes: health check and guest homepage
+  publicRoutes: ["/api/health", "/"],
   // Redirect to portal for sign-in (centralized auth)
   signInUrl: `${portalUrl}/sign-in`,
 });
