+++ apps/content-generator/package.json
@@ -17,6 +17,7 @@
     "@asafarim/db": "workspace:*",
     "@asafarim/shared-i18n": "workspace:*",
     "@asafarim/ui": "workspace:*",
+    "lucide-react": "^0.462.0",
     "next": "^16.2.3",
     "next-auth": "^5.0.0-beta.28",
     "react": "^19.0.0",
